requirements.txt
# Core
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
pyyaml>=6.0
tqdm>=4.66

# Hugging Face datasets (streaming)
datasets>=2.19
huggingface_hub>=0.24

# Embeddings / models
torch>=2.1
transformers>=4.41
sentence-transformers>=2.7

# Viz / optional
matplotlib>=3.8
umap-learn>=0.5
orjson==2.0.7


huggingface_hub>=0.22.0
pyarrow>=16.0.0
